requirements.txt
asgiref==3.7.2
boto3==1.28.12
botocore==1.31.12
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cloudinary==1.33.0
cryptography==41.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-ses==3.5.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
gunicorn==21.2.0
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
packaging==23.1
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
whitenoise==6.5.0
